rewrite more commands
